backend/requirements.txt
django==4.1
djangorestframework==3.13.0
Pillow==9.2.0
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-ckeditor==6.5.1